Ultimos cambios

+++ src/main/java/pe/gob/senamhi/contratacionesapp/repositories/IRoleRepository.java
@@ -2,7 +2,7 @@ package pe.gob.senamhi.contratacionesapp.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.gob.senamhi.contratacionesapp.entities.Role;
 
-public interface IRoleRepository extends JpaRepository<Role, Integer> {
+public interface IRoleRepository extends JpaRepository<Role, Long> {
 	Role findByRole(String role);
 
 }
